move MixinConfigMergingTransformer and plugin definitions to buildSrc
