+++ build.gradle.kts
@@ -1,9 +1,5 @@
 @file:Suppress("UnstableApiUsage")
 import com.github.jengelman.gradle.plugins.shadow.tasks.ShadowJar
-import com.github.jengelman.gradle.plugins.shadow.transformers.Transformer
-import com.github.jengelman.gradle.plugins.shadow.transformers.TransformerContext
-import groovy.json.JsonOutput
-import groovy.json.JsonSlurper
 import kotlinx.serialization.encodeToString
 import me.modmuss50.mpp.HttpUtils
 import me.modmuss50.mpp.PublishModTask
@@ -15,8 +11,6 @@ import okhttp3.MultipartBody
 import okhttp3.Request
 import okhttp3.RequestBody.Companion.asRequestBody
 import okhttp3.internal.immutableListOf
-import org.apache.tools.zip.ZipEntry
-import org.apache.tools.zip.ZipOutputStream
 import org.objectweb.asm.ClassReader
 import org.objectweb.asm.ClassWriter
 import org.objectweb.asm.tree.ClassNode
@@ -332,75 +326,6 @@ dependencies {
 
 tasks.jar {
 	enabled = false
-}
-
-class MixinConfigMergingTransformer : Transformer {
-	private val JSON = JsonSlurper()
-
-	@Input lateinit var modId: String
-	@Input lateinit var packageName: String
-	@Input lateinit var mixinPlugin: String
-
-	override fun getName(): String {
-		return "MixinConfigMergingTransformer"
-	}
-
-	override fun canTransformResource(element: FileTreeElement?): Boolean {
-		return element != null && (element.name.endsWith(".mixins.json") || element.name.endsWith("-refmap.json"))
-	}
-
-	private var transformed = false
-	
-	private var mixins = ArrayList<String>()
-	private var refMaps = HashMap<String, Map<String, String>>()
-
-	override fun transform(context: TransformerContext?) {
-		if (context == null)
-			return
-
-		this.transformed = true
-
-		val parsed = JSON.parse(context.`is`) as Map<*, *>
-		if (parsed.contains("client")) {
-			@Suppress("UNCHECKED_CAST")
-			mixins.addAll(parsed["client"] as List<String>)
-		} else {
-			@Suppress("UNCHECKED_CAST")
-			refMaps.putAll(parsed["mappings"] as Map<String, Map<String, String>>)
-		}
-	}
-
-	override fun hasTransformedResource(): Boolean {
-		return transformed
-	}
-
-	override fun modifyOutputStream(os: ZipOutputStream?, preserveFileTimestamps: Boolean) {
-		val mixinConfigEntry = ZipEntry("${modId}.mixins.json")
-		os!!.putNextEntry(mixinConfigEntry)
-		os.write(JsonOutput.prettyPrint(JsonOutput.toJson(mapOf(
-			"required" to true,
-			"minVersion" to "0.8",
-			"package" to packageName,
-			"plugin" to mixinPlugin,
-			"compatibilityLevel" to "JAVA_8",
-			"client" to mixins,
-			"injectors" to mapOf(
-				"defaultRequire" to 1,
-			),
-			"refmap" to "${modId}-refmap.json",
-		))).toByteArray())
-		
-		val refMapEntry = ZipEntry("${modId}-refmap.json")
-		os.putNextEntry(refMapEntry)
-		os.write(JsonOutput.prettyPrint(JsonOutput.toJson(mapOf(
-			"mappings" to refMaps,
-		))).toByteArray())
-
-		transformed = false
-		mixins.clear()
-		refMaps.clear()
-	}
-
 }
 
 tasks.shadowJar {
