+++ settings.gradle.kts
@@ -1,32 +1,11 @@
 pluginManagement {
     repositories {
-	    mavenCentral()
-	    maven("https://maven.fabricmc.net/")
-	    maven("https://maven.glass-launcher.net/babric")
-	    maven("https://repo.legacyfabric.net/repository/legacyfabric/")
-	    maven("https://maven.wagyourtail.xyz/releases")
-	    maven("https://maven.wagyourtail.xyz/snapshots")
-        gradlePluginPortal {
-	        content {
-		        excludeGroup("org.apache.logging.log4j")
-	        }
-        }
+	    gradlePluginPortal()
     }
-}
-
-buildscript {
-	dependencies {
-		classpath("org.ow2.asm:asm-tree:9.7")
-	}
 }
 
 plugins {
 	id("org.gradle.toolchains.foojay-resolver-convention") version("0.7.0")
-	id("xyz.wagyourtail.unimined") version("1.2.3") apply(false)
-	id("com.github.gmazzo.buildconfig") version("5.2.0") apply(false)
-	id("com.github.johnrengelman.shadow") version("8.1.1") apply(false)
-	id("me.modmuss50.mod-publish-plugin") version("0.4.5") apply(false)
-	id("org.ajoberstar.grgit") version("5.2.2") apply(false)
 }
 
 rootProject.name = "zume"
